backend/requirements.txt
Flask
Flask-Cors
sentence-transformers
spaCy
pandas
scikit-learn
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
pytesseract
pillow
